backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.14.1
amqp==5.3.1
aniso8601==10.0.0
annotated-types==0.7.0
anyio==3.7.1
archspec 
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
Authlib==1.4.0
autobahn==24.4.2
Automat==25.4.16
backoff==2.2.1
bcrypt==4.3.0
billiard==3.6.4.0
bleach==6.2.0
blinker==1.9.0
boltons 
Brotli==1.1.0
build==1.3.0
cachetools==5.5.2
capstone==5.0.6
celery==5.3.4
zeep==4.2.1
xmltodict==0.13.0
django-celery-beat==2.5.0
certifi==2024.12.14
cffi==2.0.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer 
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
configparser==7.2.0
constantly==23.10.4
cryptography==46.0.1
daphne==4.1.2
dataclasses-json==0.6.7
distro 
Django==4.2.7
django-cors-headers==4.3.0
django-jazzmin==2.6.0
django-summernote==0.8.20.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docopt==0.6.2
durationpy==0.10
Exscript==2.6.28
fastapi==0.116.1
filelock==3.19.1
Flask==3.1.0
Flask-Compress==1.17
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
frozendict 
frozenlist==1.7.0
fsspec==2025.7.0
future==1.0.0
gevent==24.11.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
gunicorn==21.2.0
h11==0.16.0
hf-xet==1.1.8
hiredis==3.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.10.0
jsonpatch==1.33
jsonpointer==2.1
keystone-engine==0.9.2
kombu==5.4.2
kubernetes==33.1.0
langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.53
langchain-openai==0.1.3
langchain-text-splitters==0.0.2
langsmith==0.1.147
lxml==5.4.0
Mako==1.3.8
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
mypy_extensions==1.1.0
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.102.0
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-asgi==0.57b0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
orjson==3.11.3
overrides==7.7.0
packaging==23.2
pandas==2.1.4
paramiko==3.5.1
passlib==1.7.4
platformdirs 
pluggy 
posthog==6.7.0
prompt_toolkit==3.0.50
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.9
pulsar-client==3.8.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.22.0
pycryptodomex==3.22.0
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==25.3.0
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyserial==3.5
PySocks 
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2024.2
pyusb==1.3.1
PyYAML==6.0.3
qrcode==8.1
redis==6.4.0
regex==2025.7.34
requests 
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rsa==4.9.1
ruamel.yaml 
ruamel.yaml.clib 
sendgrid==6.11.0
service-identity==24.2.0
setuptools==75.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
starlette==0.47.3
sympy==1.14.0
tenacity==8.5.0
tika==2.6.0
tiktoken==0.11.0
tokenizers==0.21.4
tqdm==4.67.1
truststore 
Twisted==25.5.0
txaio==25.6.1
typer==0.16.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.35.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.44.0
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
zope.event==5.0
zope.interface==7.2
zstandard 
phonepe-sdk
django-filter
drf-nested-routers
Pillow
